feat: cleanup
