+++ frontend/components/shared/traces/trace-view.tsx
@@ -15,7 +15,6 @@ import CondensedTimeline from "@/components/traces/trace-view/condensed-timeline
 import LangGraphView from "@/components/traces/trace-view/lang-graph-view";
 import LangGraphViewTrigger from "@/components/traces/trace-view/lang-graph-view-trigger";
 import List from "@/components/traces/trace-view/list";
-import { ScrollContextProvider } from "@/components/traces/trace-view/scroll-context";
 import TraceViewStoreProvider, {
   type TraceViewSpan,
   type TraceViewTrace,
@@ -115,108 +114,106 @@ export const PureTraceView = ({ trace, spans, onClose }: TraceViewProps) => {
   }, []);
 
   return (
-    <ScrollContextProvider>
-      <div className="flex flex-col h-full w-full overflow-hidden">
-        {!onClose && (
-          <div className="flex flex-none items-center border-b px-4 py-3.5 gap-2">
-            <Link className="mr-2" href="/projects">
-              <Image alt="Laminar logo" src={fullLogo} width={120} height={20} />
-            </Link>
-          </div>
-        )}
-        <ResizablePanelGroup
-          id="shared-trace-horizontal"
-          orientation="horizontal"
-          className="h-full w-full overflow-hidden"
-        >
-          <ResizablePanel id="shared-trace" defaultSize="50%" className="flex flex-col h-full overflow-hidden">
-            <Header onClose={onClose} />
-            <ResizablePanelGroup id="shared-trace-panels" orientation="vertical">
-              {condensedTimelineEnabled && (
-                <>
-                  <ResizablePanel defaultSize={200} minSize={80}>
-                    <div className="border-t h-full">
-                      <CondensedTimeline />
-                    </div>
-                  </ResizablePanel>
-                  <ResizableHandle className="hover:bg-blue-400 z-10 transition-colors hover:scale-200" />
-                </>
-              )}
-              <ResizablePanel className="flex flex-col flex-1 h-full overflow-hidden relative">
-                <div
-                  className={cn(
-                    "flex items-center gap-2 pb-2  border-b box-border transition-[padding] duration-200",
-                    condensedTimelineEnabled ? "pl-2 pr-2" : "pl-2 pr-[96px]",
-                    {
-                      "pt-2": !onClose || condensedTimelineEnabled,
-                    }
-                  )}
-                >
-                  <div className="flex items-center justify-between w-full">
-                    <div className="flex items-center gap-2">
-                      <ViewDropdown />
-                      <TraceStatsShields
-                        className="min-w-0 overflow-hidden"
-                        trace={trace}
-                        spans={filteredSpansForStats}
-                      />
-                    </div>
-                    <div className="flex items-center gap-1">
-                      <Button
-                        className={cn("h-6 px-1.5 text-xs", {
-                          "border-primary text-primary": browserSession,
-                        })}
-                        variant="outline"
-                        onClick={() => setBrowserSession(!browserSession)}
-                      >
-                        <CirclePlay size={14} className="mr-1" />
-                        Media
-                      </Button>
-                      {hasLangGraph && <LangGraphViewTrigger setOpen={setLangGraph} open={langGraph} />}
-                    </div>
+    <div className="flex flex-col h-full w-full overflow-hidden">
+      {!onClose && (
+        <div className="flex flex-none items-center border-b px-4 py-3.5 gap-2">
+          <Link className="mr-2" href="/projects">
+            <Image alt="Laminar logo" src={fullLogo} width={120} height={20} />
+          </Link>
+        </div>
+      )}
+      <ResizablePanelGroup
+        id="shared-trace-horizontal"
+        orientation="horizontal"
+        className="h-full w-full overflow-hidden"
+      >
+        <ResizablePanel id="shared-trace" defaultSize="50%" className="flex flex-col h-full overflow-hidden">
+          <Header onClose={onClose} />
+          <ResizablePanelGroup id="shared-trace-panels" orientation="vertical">
+            {condensedTimelineEnabled && (
+              <>
+                <ResizablePanel defaultSize={200} minSize={80}>
+                  <div className="border-t h-full">
+                    <CondensedTimeline />
+                  </div>
+                </ResizablePanel>
+                <ResizableHandle className="hover:bg-blue-400 z-10 transition-colors hover:scale-200" />
+              </>
+            )}
+            <ResizablePanel className="flex flex-col flex-1 h-full overflow-hidden relative">
+              <div
+                className={cn(
+                  "flex items-center gap-2 pb-2  border-b box-border transition-[padding] duration-200",
+                  condensedTimelineEnabled ? "pl-2 pr-2" : "pl-2 pr-[96px]",
+                  {
+                    "pt-2": !onClose || condensedTimelineEnabled,
+                  }
+                )}
+              >
+                <div className="flex items-center justify-between w-full">
+                  <div className="flex items-center gap-2">
+                    <ViewDropdown />
+                    <TraceStatsShields
+                      className="min-w-0 overflow-hidden"
+                      trace={trace}
+                      spans={filteredSpansForStats}
+                    />
+                  </div>
+                  <div className="flex items-center gap-1">
+                    <Button
+                      className={cn("h-6 px-1.5 text-xs", {
+                        "border-primary text-primary": browserSession,
+                      })}
+                      variant="outline"
+                      onClick={() => setBrowserSession(!browserSession)}
+                    >
+                      <CirclePlay size={14} className="mr-1" />
+                      Media
+                    </Button>
+                    {hasLangGraph && <LangGraphViewTrigger setOpen={setLangGraph} open={langGraph} />}
                   </div>
                 </div>
-                {tab === "tree" && (
-                  <div className="flex flex-1 h-full overflow-hidden relative">
-                    <Tree onSpanSelect={handleSpanSelect} isShared />
-                  </div>
-                )}
-                {tab === "transcript" && (
-                  <div className="flex flex-1 h-full overflow-hidden relative">
-                    <List onSpanSelect={handleSpanSelect} isShared />
-                  </div>
-                )}
-              </ResizablePanel>
-              {browserSession && (
-                <>
-                  <ResizableHandle className="z-50" withHandle />
-                  <ResizablePanel>
-                    <SessionPlayer
-                      onClose={() => setBrowserSession(false)}
-                      hasBrowserSession={hasBrowserSession}
-                      traceId={trace.id}
-                      llmSpanIds={llmSpanIds}
-                    />
-                  </ResizablePanel>
-                </>
-              )}
-              {langGraph && hasLangGraph && <LangGraphView spans={spans} />}
-            </ResizablePanelGroup>
-          </ResizablePanel>
-          <ResizableHandle className="hover:bg-blue-400 z-10 transition-colors hover:scale-200" />
-          <ResizablePanel id="shared-span" className="flex flex-col h-full overflow-hidden">
-            {selectedSpan ? (
-              <SpanView key={selectedSpan.spanId} spanId={selectedSpan.spanId} traceId={trace.id} />
-            ) : (
-              <div className="flex flex-col items-center justify-center size-full text-muted-foreground">
-                <span className="text-xl font-medium mb-2">No span selected</span>
-                <span className="text-base">Select a span from the trace tree to view its details</span>
               </div>
+              {tab === "tree" && (
+                <div className="flex flex-1 h-full overflow-hidden relative">
+                  <Tree onSpanSelect={handleSpanSelect} isShared />
+                </div>
+              )}
+              {tab === "transcript" && (
+                <div className="flex flex-1 h-full overflow-hidden relative">
+                  <List onSpanSelect={handleSpanSelect} isShared />
+                </div>
+              )}
+            </ResizablePanel>
+            {browserSession && (
+              <>
+                <ResizableHandle className="z-50" withHandle />
+                <ResizablePanel>
+                  <SessionPlayer
+                    onClose={() => setBrowserSession(false)}
+                    hasBrowserSession={hasBrowserSession}
+                    traceId={trace.id}
+                    llmSpanIds={llmSpanIds}
+                  />
+                </ResizablePanel>
+              </>
             )}
-          </ResizablePanel>
-        </ResizablePanelGroup>
-      </div>
-    </ScrollContextProvider>
+            {langGraph && hasLangGraph && <LangGraphView spans={spans} />}
+          </ResizablePanelGroup>
+        </ResizablePanel>
+        <ResizableHandle className="hover:bg-blue-400 z-10 transition-colors hover:scale-200" />
+        <ResizablePanel id="shared-span" className="flex flex-col h-full overflow-hidden">
+          {selectedSpan ? (
+            <SpanView key={selectedSpan.spanId} spanId={selectedSpan.spanId} traceId={trace.id} />
+          ) : (
+            <div className="flex flex-col items-center justify-center size-full text-muted-foreground">
+              <span className="text-xl font-medium mb-2">No span selected</span>
+              <span className="text-base">Select a span from the trace tree to view its details</span>
+            </div>
+          )}
+        </ResizablePanel>
+      </ResizablePanelGroup>
+    </div>
   );
 };
