+++ frontend/components/traces/trace-view/list/collapsed-text-with-more.tsx
@@ -1,13 +1,10 @@
-import { useCallback, useLayoutEffect, useRef, useState } from "react";
+import { useState } from "react";
 
 import { cn } from "@/lib/utils";
 
 const MORE_TEXT = "... more";
-
-// Average chars per pixel for 13px Inter (normal weight).
-// Inter at 13px averages ~6.2px per character. Inverse ≈ 0.161 chars/px.
-// Slightly conservative to avoid overshooting 4 lines.
 const CHARS_PER_PIXEL = 0.155;
+const FALLBACK_WIDTH = 500;
 
 interface CollapsedTextWithMoreProps {
   text: string;
@@ -33,45 +30,17 @@ export function CollapsedTextWithMore({
   charsPerPixel = CHARS_PER_PIXEL,
   className,
 }: CollapsedTextWithMoreProps) {
-  const containerRef = useRef<HTMLDivElement>(null);
   const [isExpanded, setIsExpanded] = useState(false);
-  const [truncatedText, setTruncatedText] = useState<string | null>(null);
 
-  const recalculate = useCallback(() => {
-    if (isExpanded) return;
-    const el = containerRef.current;
-    if (!el) return;
-    const width = el.clientWidth;
-    if (width === 0) return;
-
-    const charsPerLine = Math.floor(width * charsPerPixel);
-    const maxChars = charsPerLine * maxLines - MORE_TEXT.length;
-    setTruncatedText(truncateText(text, maxChars));
-  }, [text, charsPerPixel, maxLines, isExpanded]);
-
-  useLayoutEffect(() => {
-    recalculate();
-  }, [recalculate]);
-
-  useLayoutEffect(() => {
-    const el = containerRef.current;
-    if (!el || isExpanded) return;
-    const observer = new ResizeObserver(recalculate);
-    observer.observe(el);
-    return () => observer.disconnect();
-  }, [recalculate, isExpanded]);
-
-  const needsTruncation = truncatedText !== null && !isExpanded;
+  const charsPerLine = Math.floor(FALLBACK_WIDTH * charsPerPixel);
+  const maxChars = charsPerLine * maxLines - MORE_TEXT.length;
+  const truncated = isExpanded ? null : truncateText(text, maxChars);
 
   return (
-    <div
-      ref={containerRef}
-      className={cn("text-sm text-secondary-foreground", className)}
-      style={{ lineHeight: `${lineHeight}px` }}
-    >
-      {needsTruncation ? (
+    <div className={cn("text-sm text-secondary-foreground", className)} style={{ lineHeight: `${lineHeight}px` }}>
+      {truncated !== null ? (
         <p>
-          {truncatedText}
+          {truncated}
           <button
             className="text-muted-foreground hover:text-primary-foreground transition-colors cursor-pointer"
             onClick={(e) => {
